+++ VAL/src/ext/hashmap.c
@@ -769,7 +769,7 @@ uint64_t hashmap_murmur(const void* data, size_t len, uint64_t seed0,
     uint64_t seed1)
 {
     (void)seed1;
-    return MM86128(data, len, seed0);
+    return MM86128(data, (int)len, (uint32_t)seed0);
 }
 
 uint64_t hashmap_xxhash3(const void* data, size_t len, uint64_t seed0,
